Remove placement validations concern
